requirements.txt
absl-py==0.6.1
APScheduler==3.5.1
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
backcall==0.1.0
bleach==3.0.2
boto3==1.9.71
botocore==1.12.71
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
colorama==0.4.1
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==2.3.1
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
docopt==0.6.2
docutils==0.14
en-core-web-sm==2.0.0
entrypoints==0.2.3
fakeredis==0.10.3
feedparser==5.2.1
Flask==1.0.2
Flask-Cors==3.0.4
future==0.16.0
gast==0.2.0
gevent==1.2.2
greenlet==0.4.15
grpcio==1.16.0
h5py==2.7.1
humanfriendly==4.17
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonpickle==0.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.1.6
kiwisolver==1.0.1
klein==17.10.0
Markdown==2.6.11
MarkupSafe==1.1.0
matplotlib==2.2.3
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
more-itertools==4.3.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==0.28.0
nb-conda==2.2.1
nb-conda-kernels==2.2.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.7.4
numpy==1.14.5
packaging==17.1
pandocfilters==1.4.2
parso==0.3.1
pathlib==1.0.1
pickleshare==0.7.5
pika==0.11.2
plac==0.9.6
pluggy==0.8.0
preshed==1.0.1
progressbar2==3.38.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
py==1.7.0
pycparser==2.19
Pygments==2.2.0
PyHamcrest==1.9.0
pykwalify==1.6.0
pymongo==3.5.1
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyreadline==2.1
PySocks==1.6.8
pytest==4.0.2
pytest-runner==4.2
python-crfsuite==0.9.6
python-dateutil==2.7.3
python-utils==2.3.0
pytz==2018.4
pywinpty==0.5.5
PyYAML==3.12
pyzmq==17.1.2
qtconsole==4.4.3
rasa-nlu==0.13.8
redis==2.10.6
regex==2017.4.5
requests==2.18.4
ruamel.yaml==0.15.37
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.11.0
sklearn-crfsuite==0.3.6
spacy==2.0.12
tabulate==0.8.2
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
thinc==6.10.3
toolz==0.9.0
tornado==5.1.1
tqdm==4.23.3
traitlets==4.3.2
Twisted==18.9.0
typing==3.6.4
tzlocal==1.5.1
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.0.1
wincertstore==0.2
wrapt==1.10.11
zope.interface==4.6.0
